fixes
:hammer: changed int for long on hist response and hit table
:hammer: deleted Result response on data

+++ app/src/main/java/com/pedrogomez/hnmob/models/api/HitResponse.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class HitResponse(
         val objectID: String,
         val author: String,
-        val created_at_i: Int,
+        val created_at_i: Long,
         val story_title: String?,
         val story_url: String?,
         val title: String?,
